# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Seu Zuca — Marketplace B2B de Materiais de Construção
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: Listar produtos do catálogo
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: supplierId
          in: query
          schema:
            type: integer
        - name: available
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lista de produtos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Obter produto por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{id}/reviews:
    get:
      operationId: getProductReviews
      tags: [products]
      summary: Avaliações do produto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lista de avaliações
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

  # SUPPLIER PRODUCTS
  /supplier/products:
    get:
      operationId: listSupplierProducts
      tags: [supplier]
      summary: Listar produtos do fornecedor
      responses:
        "200":
          description: Lista de produtos do fornecedor
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductDetail"
    post:
      operationId: createProduct
      tags: [supplier]
      summary: Criar produto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Produto criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
  /supplier/products/{id}:
    put:
      operationId: updateProduct
      tags: [supplier]
      summary: Atualizar produto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "200":
          description: Produto atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
    delete:
      operationId: deleteProduct
      tags: [supplier]
      summary: Excluir produto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Excluído
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /supplier/quotes/{id}/respond:
    post:
      operationId: respondToQuote
      tags: [supplier]
      summary: Responder cotação
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteResponseBody"
      responses:
        "201":
          description: Resposta enviada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteResponse"
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: Listar cotações do comprador
      responses:
        "200":
          description: Lista de cotações
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Criar solicitação de cotação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteBody"
      responses:
        "201":
          description: Cotação criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Obter cotação com respostas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cotação encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteWithResponses"
  /quotes/{id}/accept:
    post:
      operationId: acceptQuoteResponse
      tags: [quotes]
      summary: Aceitar resposta de cotação
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptQuoteBody"
      responses:
        "200":
          description: Cotação aceita
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # REVIEWS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: string
      required:
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # AUTH SCHEMAS
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        nome:
          type: string
        role:
          type: string
          enum: [admin, buyer, supplier]
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        cnpj:
          type: string
        razaoSocial:
          type: string
        nomeFantasia:
          type: string
        telefone:
          type: string
        ramo:
          type: string
        emailVerificado:
          type: boolean
        createdAt:
          type: string
        stripeAccountId:
          type: string
      required:
        - id
        - email
        - nome
        - role
        - status
...
    Category:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        slug:
          type: string
        descricao:
          type: string
        parentId:
          type: integer
        unidadeMedida:
          type: string
        ativo:
          type: boolean
        ordem:
          type: integer
        minimumRule:
          $ref: "#/components/schemas/CategoryMinimumRule"
        children:
          type: array
          items:
            $ref: "#/components/schemas/Category"
      required:
        - id
        - nome
        - slug
        - ativo
...
    CategoryMinimumRule:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        quantidadeMinima:
          type: integer
        multiplo:
          type: integer
        ativo:
          type: boolean
        categoryName:
          type: string
      required:
        - id
        - categoryId
        - quantidadeMinima
        - multiplo
        - ativo
...
    Product:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        slug:
          type: string
        descricao:
          type: string
        sku:
          type: string
        preco:
          type: number
        unidadeMedida:
          type: string
        estoque:
          type: integer
        disponivel:
          type: boolean
        categoryId:
          type: integer
        categoryName:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: string
        imagemPrincipal:
          type: string
        mediaAvaliacao:
          type: number
        totalAvaliacoes:
          type: integer
        prazoFrete:
          type: integer
        quantidadeMinima:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - nome
        - slug
        - disponivel
        - categoryId
        - supplierId
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            imagens:
              type: array
              items:
                type: string
            regioesAtendidas:
              type: array
              items:
                type: string
            category:
              $ref: "#/components/schemas/Category"
            supplier:
              $ref: "#/components/schemas/User"
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - products
        - total
        - page
        - totalPages
    CreateProductBody:
      type: object
      properties:
        nome:
          type: string
        descricao:
          type: string
        sku:
          type: string
        preco:
          type: number
        unidadeMedida:
          type: string
        estoque:
          type: integer
        categoryId:
          type: integer
        imagens:
          type: array
          items:
            type: string
        prazoFrete:
          type: integer
        regioesAtendidas:
          type: array
          items:
            type: string
        alertaEstoque:
          type: integer
      required:
        - nome
        - preco
        - unidadeMedida
        - estoque
        - categoryId

    # CART
...
    QuoteItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantidade:
          type: integer
      required:
        - id
        - productId
        - quantidade
    Quote:
      type: object
      properties:
        id:
          type: integer
        buyerId:
          type: integer
        buyerName:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: string
        status:
          type: string
          enum: [pendente, respondida, aceita, recusada, expirada]
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItem"
        observacoes:
          type: string
        createdAt:
          type: string
      required:
        - id
        - buyerId
        - status
        - createdAt
    QuoteWithResponses:
      allOf:
        - $ref: "#/components/schemas/Quote"
        - type: object
          properties:
            responses:
              type: array
              items:
                $ref: "#/components/schemas/QuoteResponse"
    QuoteResponse:
      type: object
      properties:
        id:
          type: integer
        quoteId:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: string
        precoTotal:
          type: number
        prazoEntrega:
          type: integer
        condicoes:
          type: string
        valorFrete:
          type: number
        validadeAte:
          type: string
        status:
          type: string
          enum: [pendente, aceita, recusada]
        createdAt:
          type: string
      required:
        - id
        - quoteId
        - supplierId
        - precoTotal
        - prazoEntrega
        - status
        - createdAt
    QuoteResponseBody:
      type: object
      properties:
        precoTotal:
          type: number
        prazoEntrega:
          type: integer
        condicoes:
          type: string
        valorFrete:
          type: number
        validadeAte:
          type: string
      required:
        - precoTotal
        - prazoEntrega
    CreateQuoteBody:
      type: object
      properties:
        supplierId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItemInput"
        observacoes:
          type: string
      required:
        - supplierId
        - items
    QuoteItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantidade:
          type: integer
      required:
        - productId
        - quantidade
    AcceptQuoteBody:
      type: object
      properties:
        responseId:
          type: integer
      required:
        - responseId

    # REVIEWS
    Review:
      type: object
      properties:
        id:
          type: integer
        buyerId:
          type: integer
        buyerName:
          type: string
        productId:
          type: integer
        productName:
          type: string
        supplierId:
          type: integer
        nota:
          type: integer
        comentario:
          type: string
        aprovada:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - buyerId
        - nota
        - comentario
        - aprovada
        - createdAt
...