# lib/api-zod/src/generated/types/quoteResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Seu Zuca — Marketplace B2B de Materiais de Construção
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteResponseStatus } from "./quoteResponseStatus";

export interface QuoteResponse {
  id: number;
  quoteId: number;
  supplierId: number;
  supplierName?: string;
  precoTotal: number;
  prazoEntrega: number;
  condicoes?: string;
  valorFrete?: number;
  validadeAte?: string;
  status: QuoteResponseStatus;
  createdAt: string;
}